fix: resolve missing draco headers in skp2xml and fix C4267 warnings

+++ src/gltflib/parsemeshdata.cpp
@@ -19,8 +19,8 @@ void ParseMeshdata::getAllBufferViewers()
     allBufferViewers.clear();
     for (auto iter = model->bufferViews.begin(); iter != model->bufferViews.end(); ++iter)
     {
-        int start = iter->byteOffset;
-        int count = iter->byteLength;
+        size_t start = iter->byteOffset;
+        size_t count = iter->byteLength;
         std::vector<uchar> bufferViewer(model->buffers[iter->buffer].data.begin() + start, model->buffers[iter->buffer].data.begin() + start + count);
         allBufferViewers.emplace_back(bufferViewer);
     }
@@ -53,10 +53,10 @@ void ParseMeshdata::getAllMeshes()
                 int byteLength        = tinygltf::GetComponentSizeInBytes(componentType);
                 size_t type           = model->accessors[position].type;
                 size_t typeSize       = tinygltf::GetNumComponentsInType(type);
-                int byteStride =
+                size_t byteStride =
                     model->bufferViews[bufferViewerIndex].byteStride == 0 ? typeSize * byteLength : model->bufferViews[bufferViewerIndex].byteStride;
-                int start                         = model->accessors[position].byteOffset;
-                int end                           = start + model->accessors[position].count * byteStride;
+                size_t start                         = model->accessors[position].byteOffset;
+                size_t end                           = start + model->accessors[position].count * byteStride;
                 std::vector<uchar> &bufferViewers = allBufferViewers[bufferViewerIndex];
                 std::vector<uchar> bufferViewer(bufferViewers.begin() + start, bufferViewers.begin() + end);
                 for (size_t i = 0; i < bufferViewer.size(); i += byteStride)
@@ -87,10 +87,10 @@ void ParseMeshdata::getAllMeshes()
                 int byteLength        = tinygltf::GetComponentSizeInBytes(componentType);
                 size_t type           = model->accessors[color].type;
                 size_t typeSize       = tinygltf::GetNumComponentsInType(type);
-                int byteStride =
+                size_t byteStride =
                     model->bufferViews[bufferViewerIndex].byteStride == 0 ? typeSize * byteLength : model->bufferViews[bufferViewerIndex].byteStride;
-                int start                         = model->accessors[color].byteOffset;
-                int end                           = start + model->accessors[color].count * byteStride;
+                size_t start                         = model->accessors[color].byteOffset;
+                size_t end                           = start + model->accessors[color].count * byteStride;
                 std::vector<uchar> &bufferViewers = allBufferViewers[bufferViewerIndex];
                 std::vector<uchar> bufferViewer(bufferViewers.begin() + start, bufferViewers.begin() + end);
                 for (size_t i = 0; i < bufferViewer.size(); i += byteStride)
@@ -121,10 +121,10 @@ void ParseMeshdata::getAllMeshes()
                 int byteLength        = tinygltf::GetComponentSizeInBytes(componentType);
                 size_t type           = model->accessors[normal].type;
                 size_t typeSize       = tinygltf::GetNumComponentsInType(type);
-                int byteStride =
+                size_t byteStride =
                     model->bufferViews[bufferViewerIndex].byteStride == 0 ? typeSize * byteLength : model->bufferViews[bufferViewerIndex].byteStride;
-                int start                         = model->accessors[normal].byteOffset;
-                int end                           = start + model->accessors[normal].count * byteStride;
+                size_t start                         = model->accessors[normal].byteOffset;
+                size_t end                           = start + model->accessors[normal].count * byteStride;
                 std::vector<uchar> &bufferViewers = allBufferViewers[bufferViewerIndex];
                 std::vector<uchar> bufferViewer(bufferViewers.begin() + start, bufferViewers.begin() + end);
                 for (size_t i = 0; i < bufferViewer.size();)
@@ -151,14 +151,14 @@ void ParseMeshdata::getAllMeshes()
             {
                 int normal            = attributes["TEXCOORD_0"];
                 int bufferViewerIndex = model->accessors[normal].bufferView;
-                uint componentType    = model->accessors[normal].componentType;
+                size_t componentType  = model->accessors[normal].componentType;
                 int byteLength        = tinygltf::GetComponentSizeInBytes(componentType);
-                uint type             = model->accessors[normal].type;
-                uint typeSize         = tinygltf::GetNumComponentsInType(type);
-                int byteStride =
+                size_t type             = model->accessors[normal].type;
+                size_t typeSize         = tinygltf::GetNumComponentsInType(type);
+                size_t byteStride =
                     model->bufferViews[bufferViewerIndex].byteStride == 0 ? typeSize * byteLength : model->bufferViews[bufferViewerIndex].byteStride;
-                int start                         = model->accessors[normal].byteOffset;
-                int end                           = start + model->accessors[normal].count * byteStride;
+                size_t start                         = model->accessors[normal].byteOffset;
+                size_t end                           = start + model->accessors[normal].count * byteStride;
                 std::vector<uchar> &bufferViewers = allBufferViewers[bufferViewerIndex];
                 std::vector<uchar> bufferViewer(bufferViewers.begin() + start, bufferViewers.begin() + end);
                 for (size_t i = 0; i < bufferViewer.size();)
